dcfg: Take into account MEM_PLL_CFG_SHIFT for ddr frequency
The base ddr clock frequency is 1/4 the speed of the final
ddr clock frequency. By default the RCW config is setting
the CFG_SHIFT to be a 1/4 divider of the memory speed,
i.e.  3200 / 4 (800MHz).  The parent DDRCLK is 100MHz so PLL_RAT
needs to be rounded to 100MHz. However using a divider of / 3 for
a lower speed always us to match industry standard ddr4 speeds
2666 and 2933 (2000 / 3 * 4 = 2666.67)

This patch takes into account the PLL_CFG_SHIFT divider so these
speeds can be configured in the RCW and then updates the helper
function that reports ddr_clk_freq to properly multiply the
clock fed into the DDRC * 4 to properly reflect the actual MTs
that the memory is being configured for.

Signed-off-by: Jon Nettleton <jon@solid-run.com>